+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 1.1.0 - 2020-08-04
+- Added query scope `aActiveSuspensions` to get all active suspensions
+- Added query scope `noneActiveSuspensions` to get all non-active suspensions
 ## 1.0.1 - 2020-07-29
 - fixed issue with model suspension check for instances 
 when model is not currently suspended